+++ ecommerce-inventory-ui/src/app/core/models/index.ts
@@ -6,5 +6,6 @@ export * from './order';
 export * from './unprepared-item-dto';
 export * from './api-response';
 export * from './marketplace-stats';
+export * from './price-history';
 export * from './reporting';
 export * from './reporting.models';